修正2 3 4 項目

+++ server.js
@@ -4,6 +4,8 @@ const mongoose = require('mongoose');
 const headers = require('./headers');
 const handleSuccess = require('./scuessHandle');
 const handleError = require('./errorHandle');
+
+// model 載入
 const Posts = require('./models/posts');
 
 dotenv.config({path: './config.env'});
@@ -14,7 +16,7 @@ const DB = process.env.DATABASE.replace(
 );
 
 mongoose
-//   .connect('mongodb://localhost:27017/hotel') // 本機端
+  // .connect('mongodb://localhost:27017/hotel') // 本機端
   .connect(DB)
   .then(() => console.log('資料庫連接成功'));
 
@@ -51,16 +53,41 @@ const requestListener = async (req, res) => {
     const allPosts = await Posts.find();
     handleSuccess(res, allPosts);
   } else if (req.url.startsWith('/posts/') && req.method === 'DELETE') {
-    const id = req.url.split('/').pop();
-    const posts = await Posts.findByIdAndDelete(id);
-    const allPosts = await Posts.find();
-    posts ? handleSuccess(res, allPosts) : handleError(res);
+    req.on('end', async () => {
+      try{
+        const id = req.url.split('/').pop();
+        console.log(id.length)
+        if(id.length <= 23) {
+          handleError(res, '', '無此id');
+          return;
+        }
+        const posts = await Posts.findByIdAndDelete(id);
+        console.log(posts)
+        const allPosts = await Posts.find();
+        if(posts == null) {
+          handleSuccess(res, allPosts, '尚未找到此筆資料')
+        } else {
+          posts ? handleSuccess(res, allPosts, '已刪除此筆資料') : handleError(res);
+        }
+      } catch(e){
+        console.log(e)
+      }
+      
+    })
   }  else if (req.url.startsWith('/posts/') && req.method === 'PATCH') {
     req.on('end', async () => {
         const id = req.url.split('/').pop();
         const data = JSON.parse(body);
+        const arg = ['name', 'tags', 'type', 'content']
+        const result = await arg.filter(key => data[key] == '' || data[key] == undefined)
+        if(result.length > 0) {
+          handleError(res);
+          return;
+        }
         const posts = await Posts.findByIdAndUpdate(id, data, { new: true});
-        handleSuccess(res, posts)
+        if(posts == null) handleError(res)
+        else handleSuccess(res, posts)
+        
     })
   } else if (req.method === 'OPTIONS') {
     res.writeHead(200, headers);
